# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "學生選課系統",
        "description": "Laravel 10 學生選課系統 API 文件",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravel10-course.test/",
            "description": "本地開發環境"
        }
    ],
    "paths": {
...
        "/api/courses": {
            "get": {
                "tags": [
                    "課程"
                ],
                "summary": "課程列表",
                "operationId": "02eeda67042dba2d22a517f26fb8eaf1",
                "responses": {
                    "200": {
                        "description": "課程列表",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "中級英文"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "中級英文課程，適合初學者"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "0900"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "1030"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "中級英文",
                                                    "description": "中級英文課程，適合初學者",
                                                    "start_time": "0900",
                                                    "end_time": "1030"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "程式設計入門",
                                                    "description": "學習基礎程式設計概念",
                                                    "start_time": "1400",
                                                    "end_time": "1530"
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "課程"
                ],
                "summary": "新增課程",
                "operationId": "a4d6b7e7be59571034428ffbb86aaf0a",
                "requestBody": {
                    "description": "課程資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "start_time",
                                    "end_time",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "課程名稱",
                                        "type": "string",
                                        "example": "程式設計入門"
                                    },
                                    "description": {
                                        "description": "課程描述",
                                        "type": "string",
                                        "example": "學習基礎程式設計概念"
                                    },
                                    "start_time": {
                                        "description": "開始時間（24小時制）",
                                        "type": "string",
                                        "example": "0900"
                                    },
                                    "end_time": {
                                        "description": "結束時間（24小時制）",
                                        "type": "string",
                                        "example": "1030"
                                    },
                                    "user_id": {
                                        "description": "教師ID",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "課程新增成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "程式設計入門"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "學習基礎程式設計概念"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "example": "0900"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "example": "1030"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "新增課程成功"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "課程名稱是必填的"
                                                    }
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "課程描述是必填的"
                                                    }
                                                },
                                                "start_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "開始時間是必填的"
                                                    }
                                                },
                                                "end_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "結束時間是必填的"
                                                    }
                                                },
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "老師是必填的"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "新增課程失敗"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/courses/{course}": {
            "put": {
                "tags": [
                    "課程"
                ],
                "summary": "更新課程",
                "operationId": "5f1701987d5321072267c5f888fd05af",
                "parameters": [
                    {
                        "name": "course",
                        "in": "path",
                        "description": "課程 ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "課程資料",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "start_time",
                                    "end_time",
                                    "user_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "課程名稱",
                                        "type": "string",
                                        "example": "中級英文"
                                    },
                                    "description": {
                                        "description": "課程描述",
                                        "type": "string",
                                        "example": "中級英文課程，適合初學者"
                                    },
                                    "start_time": {
                                        "description": "開始時間（24小時制）",
                                        "type": "string",
                                        "example": "0900"
                                    },
                                    "end_time": {
                                        "description": "結束時間（24小時制）",
                                        "type": "string",
                                        "example": "1030"
                                    },
                                    "user_id": {
                                        "description": "教師ID",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "課程更新成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "中級英文"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "中級英文課程，適合初學者"
                                                },
                                                "start_time": {
                                                    "type": "string",
                                                    "example": "0900"
                                                },
                                                "end_time": {
                                                    "type": "string",
                                                    "example": "1030"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "更新課程成功"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "驗證失敗",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "課程名稱是必填的"
                                                    }
                                                },
                                                "description": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "課程描述是必填的"
                                                    }
                                                },
                                                "start_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "開始時間是必填的"
                                                    }
                                                },
                                                "end_time": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "結束時間是必填的"
                                                    }
                                                },
                                                "user_id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "老師是必填的"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "更新課程失敗"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}